Update libfwps to version 20220122 / rev 4 via SR 1038477
https://build.opensuse.org/request/show/1038477
by user jengelh + dimstar_suse
delete system-libs.patch

+++ packages/libf/libfwps/libfwps.spec
@@ -27,11 +27,8 @@ URL:            https://github.com/libyal/libfwps
 Source:         https://github.com/libyal/libfwps/releases/download/%version/libfwps-alpha-%version.tar.gz
 Source2:        https://github.com/libyal/libfwps/releases/download/%version/libfwps-alpha-%version.tar.gz.asc
 Source9:        %name.keyring
-Patch1:         system-libs.patch
 BuildRequires:  %{python_module devel}
 BuildRequires:  c_compiler
-BuildRequires:  gettext-tools >= 0.18.1
-BuildRequires:  libtool
 BuildRequires:  pkg-config
 BuildRequires:  python-rpm-macros
 BuildRequires:  pkgconfig(libcdata) >= 20220115
@@ -44,6 +41,7 @@ BuildRequires:  pkgconfig(libfguid) >= 20220113
 BuildRequires:  pkgconfig(libfole) >= 20220115
 BuildRequires:  pkgconfig(libuna) >= 20220102
 %python_subpackages
+# Various notes: https://en.opensuse.org/libyal
 
 %description
 libfwps is a library for Windows Property Store data types.
@@ -74,22 +72,19 @@ applications that want to make use of libfwps.
 %autosetup -p1
 
 %build
-autoreconf -fi
-# OOT builds are presently broken, so we have to install
-# within each python iteration now, not in %%install.
 %{python_expand #
-# see libcdata for version-sc
 echo "V_%version { global: *; };" >v.sym
 %configure --disable-static \
 	--enable-python PYTHON_VERSION="%{$python_bin_suffix}" \
 	LDFLAGS="-Wl,--version-script=$PWD/v.sym"
+grep ' '' ''local' config.log && exit 1
 %make_build
 %make_install DESTDIR="%_builddir/rt"
 %make_build clean
 }
 
 %install
-mv %_builddir/rt/* %buildroot/
+mv "%_builddir/rt"/* %buildroot/
 find "%buildroot" -type f -name "*.la" -delete -print
 
 %post   -n %lname -p /sbin/ldconfig
